Fully implement non-local variable access
Uses a scheme with a side-table (the "display"), which has one entry for each
captured variable in the program. Each entry points to the location of the
variable in the most-recent-caller. On entry to a function whose locals are
captured, it stores copies of the relevent entries from the display and
updates the display with pointers to its own locals. Later functions that wish
to access the nonlocal look up the pointer to it from the display.