don't use enum type as va_start argument
Unfortunately, due to completely reasonable restrictions, enum types
cannot be used as the "anchor" for varargs in C. In C++, we could set
the size of the enum explicitly and not have this issue. This would only
arise on platforms where enums are not always represented as an int. For
example, ARM. Thanks for the warning, clang! (Eat a dick, gcc!)